+++ src/meltano/core/transform_add_service.py
@@ -2,18 +2,35 @@ import os
 import json
 import yaml
 import logging
+from pathlib import Path
 
 from .project import Project
-from .plugin import Plugin, PluginInstall
+from .config_service import ConfigService
+from .plugin.settings_service import PluginSettingsService
+from .plugin import Plugin, PluginType, PluginInstall
+from .db import project_engine
 
 
 class TransformAddService:
     def __init__(self, project: Project):
         self.project = project
-        self.packages_file = project.root.joinpath("transform").joinpath("packages.yml")
-        self.dbt_project_file = project.root.joinpath("transform").joinpath(
-            "dbt_project.yml"
+
+        _, Session = project_engine(project)
+        session = Session()
+
+        config_service = ConfigService(project)
+        dbt_plugin = config_service.find_plugin(
+            plugin_name="dbt", plugin_type=PluginType.TRANSFORMERS
         )
+
+        plugin_settings_service = PluginSettingsService(project)
+        dbt_project_dir, _ = plugin_settings_service.get_value(
+            session, dbt_plugin, "project_dir"
+        )
+        dbt_project_path = Path(dbt_project_dir)
+
+        self.packages_file = dbt_project_path.joinpath("packages.yml")
+        self.dbt_project_file = dbt_project_path.joinpath("dbt_project.yml")
 
     def add_to_packages(self, plugin: Plugin):
         if not os.path.exists(self.packages_file):