+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - [!1664](https://gitlab.com/meltano/meltano/-/merge_requests/1664) Set `MELTANO_EXTRACTOR_NAME`, `MELTANO_EXTRACTOR_NAMESPACE`, and `MELTANO_EXTRACT_{SETTING...}` environment variables when invoking loader
 - [!1664](https://gitlab.com/meltano/meltano/-/merge_requests/1664) Don't pass environment variables with "None" values to plugins if variables were unset
 - [!1664](https://gitlab.com/meltano/meltano/-/merge_requests/1664) Set `target-postgres` and `target-snowflake` schema based on `MELTANO_EXTRACTOR_NAMESPACE` env var
+- [!1664](https://gitlab.com/meltano/meltano/-/merge_requests/1664) Set `dbt` project dir based on `DBT_PROJECT_DIR` env var, defaulting to `$MELTANO_PROJECT_ROOT/transform`
 
 ### Fixes